requirements.txt
python-telegram-bot==13.15
requests
googletrans==4.0.0-rc1